backend/requirements.txt
# FastAPI 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# 定时任务
apscheduler>=3.10.4

# HTTP 请求
httpx>=0.26.0
aiohttp>=3.9.1

# RSS 解析
feedparser>=6.0.10

# 数据处理
pandas>=2.2.0
numpy>=2.0.0

# AI 服务
anthropic>=0.18.1
openai>=1.12.0

# 语音合成
edge-tts>=6.1.9
pydub>=0.25.1

# 股票数据
akshare>=1.12.25

# 配置管理
python-dotenv>=1.0.1
pydantic-settings>=2.1.0

# 工具
python-dateutil>=2.8.2
beautifulsoup4>=4.12.3
lxml>=5.1.0
markdown>=3.5.2
